requirements.txt
numpy~=1.26
pydantic>=2.11.5
pytest-mock>=3.14.1
ray[rllib]==2.46.0
pettingzoo[classic]
supersuit
wandb
gymnasium==1.0.0
